+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This changelog follows the patterns described here: <https://keepachangelog.com/
 
 Subheadings to categorize changes are `added, changed, deprecated, removed, fixed, security`.
 
-## Unreleased
+## 1.1.0
 
 ### added